+++ packages/project-access/src/project/info.ts
@@ -1,11 +1,102 @@
 import type { Editor } from 'mem-fs-editor';
 import { join } from 'path';
-import { FileName } from '../constants';
+import { DirName, FileName } from '../constants';
 import { fileExists, findFilesByExtension, readJSON } from '../file';
-import type { Package, AppProgrammingLanguage, AppType, ProjectType, AllAppResults } from '../types';
+import type {
+    AllAppResults,
+    ApplicationStructure,
+    AppProgrammingLanguage,
+    AppType,
+    Manifest,
+    Package,
+    Project,
+    ProjectType
+} from '../types';
 import { getCapProjectType } from './cap';
-import { getWebappPath } from './ui5-config';
+import { getI18nPropertiesPaths } from './i18n/i18n';
 import { findFioriArtifacts } from './search';
+import { getMainService, getServicesAndAnnotations } from './service';
+import { getWebappPath } from './ui5-config';
+
+/**
+ * Returns the project structure for a given Fiori project.
+ *
+ * @param root - project root folder
+ * @returns - project structure with project info like project type, apps, root folder
+ */
+export async function getProject(root: string): Promise<Project> {
+    const checkCapType = await getCapProjectType(root);
+    const projectType = checkCapType ?? 'EDMXBackend';
+    const packageJson = await readJSON<Package>(join(root, FileName.Package));
+    const appFolders = getAppFolders(packageJson);
+    const apps = await getApps(root, appFolders);
+    return {
+        root,
+        projectType,
+        apps
+    };
+}
+
+/**
+ * Returns the application folders from sapux flag of the package.json. For single app
+ * projects, this is just an array with one empty string. For CAP projects, this is an
+ * array of operating system specific relative paths to the apps.
+ *
+ * @param packageJson - parsed package.json
+ * @returns - array of operating specific application folders
+ */
+function getAppFolders(packageJson: Package): string[] {
+    return Array.isArray(packageJson.sapux)
+        ? packageJson.sapux.map((appFolder) => join(...appFolder.split(/[/\\]/)))
+        : [''];
+}
+
+/**
+ * Get the application structure for each application in the project.
+ *
+ * @param root - project root folder
+ * @param appFolders - array of relative application folders
+ * @returns - map of application structures
+ */
+async function getApps(root: string, appFolders: string[]): Promise<{ [index: string]: ApplicationStructure }> {
+    const apps: { [index: string]: ApplicationStructure } = {};
+    for (const appFolder of appFolders) {
+        const applicationStructure = await getApplicationStructure(root, appFolder);
+        if (applicationStructure) {
+            apps[appFolder] = applicationStructure;
+        }
+    }
+    return apps;
+}
+
+/**
+ * Get the application structure for a given application.
+ *
+ * @param root - project root folder
+ * @param appFolder - relative application folder
+ * @returns - application structure with application info like manifest, changes, main service, services, annotations
+ */
+async function getApplicationStructure(root: string, appFolder: string): Promise<ApplicationStructure | undefined> {
+    const appRoot = join(root, appFolder);
+    const absoluteWebappPath = await getWebappPath(appRoot);
+    const manifest = join(absoluteWebappPath, FileName.Manifest);
+    if (!(await fileExists(manifest))) {
+        return undefined;
+    }
+    const manifestObject = await readJSON<Manifest>(manifest);
+    const changes = join(absoluteWebappPath, DirName.Changes);
+    const i18n = await getI18nPropertiesPaths(manifest, manifestObject);
+    const mainService = getMainService(manifestObject);
+    const services = await getServicesAndAnnotations(manifest, manifestObject);
+    return {
+        appRoot,
+        manifest,
+        changes,
+        i18n,
+        mainService,
+        services
+    };
+}
 
 /**
  * Get the used programming language of an application.